chore: moved kolektivo notifications from home to goldtoken
