+++ src/goldToken/reducer.ts
@@ -4,12 +4,14 @@ export interface State {
   balance: string | null
   lastFetch: number | null
   educationCompleted: boolean
+  kolektivoNotifications: { [key: string]: boolean }
 }
 
 export const initialState = {
   balance: null,
   lastFetch: null,
   educationCompleted: false,
+  kolektivoNotifications: { cicoPrompt: true },
 }
 
 export function reducer(state: State | undefined = initialState, action: ActionTypes): State {
@@ -24,6 +26,14 @@ export function reducer(state: State | undefined = initialState, action: ActionT
       return {
         ...state,
         educationCompleted: action.educationCompleted,
+      }
+    case Actions.SET_CICO_COMPLETED:
+      return {
+        ...state,
+        kolektivoNotifications: {
+          ...state.kolektivoNotifications,
+          cicoPrompt: action.cicoPrompt,
+        },
       }
     default:
       return state
